module: implement the "module" exports condition
The "module" exports condition has been used by bundlers in
the wild to support require(esm) in bundles targeting the
Node.js environment. This patch implements the same condition
for the native require(esm) implementation in Node.js to match
bundler behaviors in the ecosystem.

The tests added in this patch match the output from webpack
bundles.

Refs: https://gist.github.com/sokra/e032a0f17c1721c71cfced6f14516c62